Refactor signup into a single file
