--- src/components/account/SignUp/constants/defaultState.ts
@@ -1,8 +1,0 @@
-import { SignUpFormFieldsType } from "../types";
-
-export const defaultSignUpFormFieldsState: SignUpFormFieldsType = {
-  name: "",
-  email: "",
-  password: "",
-  acceptsEmail: false,
-};
