--- src/components/account/SignUp/hooks/useSignUp/types.ts
@@ -1,5 +1,0 @@
-import { SignUpFormFieldsType } from "../../types";
-
-export type SignUpActionsType = {
-  signUp: (data: SignUpFormFieldsType) => void;
-};
