--- src/components/account/SignUp/types.ts
@@ -1,6 +1,0 @@
-export type SignUpFormFieldsType = {
-  name: string;
-  email: string;
-  password: string;
-  acceptsEmail?: boolean;
-};
